pyproject.toml
[tool.poetry]
name = "chorebot"
version = "0.1.0"
description = "Household chore organization with Trello."
authors = ["Eric Smith <eric@esmithy.net>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.6"
python-dateutil = "^2.8"
pytz = "^2019.1"
requests = "^2.22"
tzlocal = "^1.5"

[tool.poetry.dev-dependencies]
pytest = "^4.6"
flake8 = "^3.7"
pre-commit = "^1.17"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79
target_version = ['py36']
skip-string-normalization = true
